wip
	modified:   Makefile.common
	modified:   include/config.h
	modified:   include/csx_data.h
	modified:   include/csx_mem.h
	modified:   include/csx_mmio.h
	modified:   include/csx_statistics.h
	modified:   include/csx_test_arm_inst.h
	new file:   include/csx_test_arm_inst_sop.h
	modified:   include/soc_core.h
	modified:   include/soc_core_utility.h
	modified:   include/soc_mmio.h
	modified:   include/soc_mmio_omap.h
	modified:   include/soc_omap_5912.h
	modified:   source/csx.c
	modified:   source/csx_data.c
	modified:   source/csx_mem.c
	modified:   source/csx_mmio.c
	modified:   source/csx_soc.c
	modified:   source/csx_test.c
	modified:   source/csx_test_arm.c
	modified:   source/csx_test_arm_inst.c
	modified:   source/csx_test_thumb.c
	modified:   source/csx_test_thumb_asm.c
	modified:   source/main.c
	modified:   source/soc_core_arm.c
	modified:   source/soc_core_disasm.c
	modified:   source/soc_mmio_mpu_mmc.c
	modified:   source/soc_mmio_ocp.c
	modified:   source/soc_mmio_os_timer.c
	modified:   source/soc_mmio_uart.c
	modified:   source/soc_mmu.c

+++ include/soc_core_utility.h
@@ -15,7 +15,7 @@
 
 #define CYCLE core->csx->cycle
 
-static inline int _check_sbo(uint32_t opcode, uint8_t msb, uint8_t lsb, uint32_t *test, uint32_t *result)
+static inline int _check_sbo(uint32_t opcode, uint msb, uint lsb, uint32_t *test, uint32_t *result)
 {
 	uint32_t set_bit_count = (msb + 1) - lsb;
 	uint32_t set_bit_mask = (1 << set_bit_count) - 1;
@@ -34,7 +34,7 @@ static inline int _check_sbo(uint32_t opcode, uint8_t msb, uint8_t lsb, uint32_t
 	return(ttest != rresult);
 }
 
-static inline int _check_sbz(uint32_t opcode, uint8_t msb, uint8_t lsb, uint32_t *test, uint32_t *result)
+static inline int _check_sbz(uint32_t opcode, uint msb, uint lsb, uint32_t *test, uint32_t *result)
 {
 	uint32_t set_bit_count = (msb + 1) - lsb;
 	uint32_t set_bit_mask = (1 << set_bit_count) - 1;
